Temp fix for autobuild: VC2008 doesn't define hypotf.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@10021 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation2/components/CCmpProjectileManager.cpp
@@ -196,7 +196,7 @@ void CCmpProjectileManager::LaunchProjectile(entity_id_t source, CFixedVector3D 
 	projectile.targetEnt = targetEnt;
 
 	CVector3D offset = projectile.target - projectile.pos;
-	float horizDistance = hypotf(offset.X, offset.Z);
+	float horizDistance = hypot(offset.X, offset.Z);	// TODO: Fix missing hypotf for VC2008
 
 	projectile.speedFactor = 1.f;
 	projectile.timeLeft = horizDistance / speed.ToFloat();
